requirements.txt
universal-ink-library>=2.0.1
python-dateutil>=2.9.0
lxml==5.2.2
pillow